app/backend/flask_app/requirements.txt
https://github.com/unbit/uwsgi/archive/uwsgi-3.0.zip#egg=uwsgi
click
itsdangerous
passlib
alembic
flasgger
flask-marshmallow
marshmallow-sqlalchemy
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
Flask-Script
flask-restful
flask_httpauth
Flask-Mail
bcrypt
python-dateutil>2.2
psycopg2
celery
redis
pendulum
flower
requests
buffered_smtp_handler
